requirements.txt
asgiref>=3.3.2
Django>=3.2.24
django-ckeditor-5==0.2.13
django-js-asset==1.2.2
django-taggit==4.0.0
django-user-visit==0.4
gunicorn==20.0.4
python-decouple
python-dotenv
pytz
sqlparse==0.3.1
ua-parser==0.10.0
user-agents==2.2.0
whitenoise==5.0.1
joblib~=1.1.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
nltk~=3.7
numpy~=1.23.0
PyPDF2~=2.4.1
textract==1.6.3
docx2txt==0.8
wordcloud>=1.9.0
scikit-learn~=1.1.1
django-widget-tweaks
django-allauth>=0.64.0
django-heroku
django-simple-captcha
requests
PyJWT
cryptography
celery==5.3.4
redis==5.0.1
django-celery-email==3.0.0
channels>=4.0
mysqlclient
